Fix #4 "alda list" output is buggy whith multiple servers
Remove the static dealerSocket and create a new one for a new Connection.

AldaRequest uses one static dealerSocket for every request. When "alda list"
finds the first server the connect method on the dealerSocket is called and
a new pipe to this endpoint is created. The message gets delivered trough this
pipe. When another server is found the connect method on the same socket is
called and a second pipe with a new endpoint is attached to the Socket.
As a result the dealerSocket has two outbound pipes to two different Servers.

The policy of a Dealer Socket is to load-balance Message to all outbound pipes.
So first "Status" Message is delivered over first pipe, then the load-balance alg
finds no other pipe and sets pointer again to first one. After that a second pipe
to the second server is created and attached. But the load-balance pointer still
addresses first pipe. Second message sent gets delivered over _first_ pipe to first
server again.

If resending this message without connecting -> creating new pipe the message will
be sent via second pipe to second server.

As a quick fix change the Method getDealerSocket that returns the "singleton Socket"
to return a new Socket. In the long run one should implement some sort of socket
managment and check if there is an exisiting socket containing a pipe/endpoint to this
host and port destination. Hint: ZContext.getSockets() returns the list of sockets.