Locate alda executable for integration tests
Check for the alda executable in the following order:

1. ALDA_EXECUTABLE environment variable
If this variable is set treat the value as the path to the alda
executable.

2. /usr/local/bin/alda
Check if this installation exists and use the executable.

3. test/resources/server/alda
Fallback to the test/resources/server directory. Include this
directory in the gitignore file. Therefore a alda executable
put in there will not be tracked/submited by git.