bugfix: bail out if download fails
IMHO, the error handling was in the wrong place. If Utils.downloadFile fails, we
need to propagate the error outward so that the calling code can handle it and
bail out.

The way we have it now, the calling code has no idea if the download was
successful, and simply proceeds as if it was successful.